server: inline tiny modules
