+++ crates/server/src/store/forwarder/mod.rs
@@ -9,9 +9,19 @@
 
 //! Forwarding resolver related types
 
-mod authority;
-mod config;
+use serde::Deserialize;
 
-pub use self::authority::ForwardAuthority;
-pub use self::authority::ForwardLookup;
-pub use self::config::ForwardConfig;
+use crate::resolver::config::{NameServerConfigGroup, ResolverOpts};
+
+mod authority;
+pub use authority::{ForwardAuthority, ForwardLookup};
+
+/// Configuration for file based zones
+#[derive(Clone, Deserialize, Debug)]
+#[serde(deny_unknown_fields)]
+pub struct ForwardConfig {
+    /// upstream name_server configurations
+    pub name_servers: NameServerConfigGroup,
+    /// Resolver options
+    pub options: Option<ResolverOpts>,
+}
