+++ crates/server/src/store/sqlite/mod.rs
@@ -7,10 +7,22 @@
 
 //! SQLite serving with Dynamic DNS and journaling support
 
-pub mod authority;
-mod config;
-pub mod persistence;
+use serde::Deserialize;
 
-pub use self::authority::SqliteAuthority;
-pub use self::config::SqliteConfig;
-pub use self::persistence::Journal;
+pub mod authority;
+pub use authority::SqliteAuthority;
+pub mod persistence;
+pub use persistence::Journal;
+
+/// Configuration for zone file for sqlite based zones
+#[derive(Deserialize, PartialEq, Eq, Debug)]
+#[serde(deny_unknown_fields)]
+pub struct SqliteConfig {
+    /// path to initial zone file
+    pub zone_file_path: String,
+    /// path to the sqlite journal file
+    pub journal_file_path: String,
+    /// Are updates allowed to this zone
+    #[serde(default)]
+    pub allow_update: bool,
+}
